chore(base): Move `BaseRoomInfo` into its own module.
This patch extracts the `BaseRoomInfo` type and its implementations
inside its own module.
